requirements.txt
streamlit
pandas
openai
google-generativeai
zhipuai
httpx
